Add a 'host' Makefile to capture all of the actions.

+++ build.sh
@@ -1,7 +1,13 @@
 #!/bin/bash
 set -eu
 
-WINE_PATH="/Applications/Wine Stable.app/Contents/Resources/wine/bin/wine"
+if [ -z "${WINE_PATH}" ]; then
+  WINE_PATH="/Applications/Wine Stable.app/Contents/Resources/wine/bin/wine"
+fi
+if [ ! -f "${WINE_PATH}" ]; then
+  echo "Could not find 'wine' on path, exiting!"
+  exit 1
+fi
 
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
 rm -f *.o
